refactored pp into abstract class for alg override
